requirements.txt
asgiref==3.7.2
django==4.2.4
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
